+++ doc/changelog.md
@@ -12,3 +12,5 @@
 
 - setNick, setJobExpiration, setcolor return a PutResponse, instead of a FullUser
 - deprecate useless adminConfigured endpoint
+
+- changes the semantics of the PrintJob.userIdentification field to use the FullUser._id instead of the FullUser.shortUser